% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{func}
\alias{func}
\title{Gives cumulative data about each player across all matches}
\usage{
func(matches, data, indices)
}
\arguments{
\item{matches}{a dataframe with 2 columns: winner and loser}

\item{data}{is the dataframe containing all the data}

\item{indices}{is a vector of row numbers in 'data' corresponding to the matches in 'matches'}
}
\value{
list(w_output, l_output),
w_output: dataframe of all the data for the winner in each match
l_output: dataframe of all the data for the loser in each match
}
\description{
Gives cumulative data about each player across all matches
}
