% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_recency_weights}
\alias{get_recency_weights}
\title{Get weights to be applied to each match based on recency}
\usage{
get_recency_weights(train_set, recency_weighting)
}
\arguments{
\item{train_set}{dataframe of training set}

\item{recency_weighting}{weight to be applied to each match}
}
\value{
vector of weights to be applied to each match
Note: if train_set does not contain a 'tourney_date' column, returns NA
}
\description{
Get weights to be applied to each match based on recency
}
