% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_}
\alias{predict_}
\title{Predicts the probability that player1 will beat player2}
\usage{
predict_(player1, player2, df_coeff)
}
\arguments{
\item{player1}{name of 1st player}

\item{player2}{name of 2nd player}

\item{df_coeff}{dataframe of coefficients}
}
\value{
: vector of prediction probabilities [1] and boolean of whether player1 won [2]
}
\description{
Predicts the probability that player1 will beat player2
}
