% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_matches}
\alias{predict_matches}
\title{Predicts the probability that player1 will beat player2 for each match in test_set}
\usage{
predict_matches(test_set, names, model)
}
\arguments{
\item{test_set}{dataframe of test set}

\item{names}{vector of names of players in test set}

\item{model}{BTm model}
}
\value{
dataframe of predictions for each match in test_set
}
\description{
Predicts the probability that player1 will beat player2 for each match in test_set
}
