% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{TennisTidyr}
\alias{TennisTidyr}
\title{Get data and unique player names from year range}
\usage{
TennisTidyr(startyear, endyear, testset, numgames)
}
\arguments{
\item{startyear}{Year to start from}

\item{endyear}{Last year to include}

\item{testset}{Year of the test set}

\item{numgames}{Minimum number of games per player}
}
\value{
2 datasets, accessible by $train and $test containing the
training set and test set, and $main_names, the unique players.
}
\description{
Get data and unique player names from year range
}
