% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{time_weighting}
\alias{time_weighting}
\title{Return the weighting parameter for a given time}
\usage{
time_weighting(t, weight)
}
\arguments{
\item{t}{time since match}

\item{weight}{weight to be applied}
}
\value{
weight to be applied to match
}
\description{
Return the weighting parameter for a given time
}
